Master

+++ Shop/Shop.py
@@ -1,18 +1,30 @@
 ﻿import os
-import json
+import json 
+from random import randint
 import telebot
 from telebot import types
 
 path = os.path.dirname(__file__)
-with open(os.path.join(path, 'settings.json'), 'r', encoding='utf-8-sig') as f:
+with open(os.path.join(path, 'settings.json'), 'r', encoding='cp1251') as f:
     settings = json.load(f)
 
 with open(os.path.join(path, 'ShopItems.json'), 'r', encoding='utf-8-sig') as f:    
     shopitems = json.load(f)
 
+with open(os.path.join(path, 'customers.json'), 'r', encoding='utf-8-sig') as f:    
+    customers = json.load(f)
+
+
 bot = telebot.TeleBot(settings['token'])
 
 user_states = {}  # Словарь для хранения состояний пользователей
+
+dataname = {
+"name": "Название",
+"price": "Цена",
+"description": "Описание",
+"image": "Изображение"               
+}
 
 try:
     me = bot.get_me()
@@ -27,17 +39,24 @@ welcomeKeyboard.add(types.InlineKeyboardButton("Профиль", callback_data="
 
 subscribeKeyboard = types.InlineKeyboardMarkup()
 subscribeKeyboard.add(types.InlineKeyboardButton("Подписаться", url="https://t.me/misterLogovo"))
-
-
+ 
+def find_product_by_id(id:int):
+    for item in shopitems:
+        if item['id'] == id:
+            return item
+    return False
 
 
 @bot.callback_query_handler(func=lambda call: True)
 def handle_callback_query(call):
+    if call.message.from_user.id == me.id:
+        return
     print(f"Получен callback от пользователя {call.from_user.id} с данными: {call.data}")
     if call.data == 'products':
         shopButtons = types.InlineKeyboardMarkup()
         for item in shopitems:
-            shopButtons.add(types.InlineKeyboardButton(item['name'] + f" ({item['price']} руб.)", callback_data=f"buy_{item['id']}"))
+            print(type(item['name']))
+            shopButtons.add(types.InlineKeyboardButton(item['name'] + f" ({str(item['price'])} руб.)", callback_data=f"buy_{int(item['id'])}"))
         shopButtons.add(types.InlineKeyboardButton("Назад", callback_data="back_to_welcome"))
         bot.delete_message(call.message.chat.id, call.message.message_id)
         bot.send_message(call.message.chat.id, "Выберите товар из списка ниже:", reply_markup=shopButtons)
@@ -54,29 +73,93 @@ def handle_callback_query(call):
                 productKeyboard = types.InlineKeyboardMarkup()
                 productKeyboard.add(types.InlineKeyboardButton("Купить", callback_data=f"confirm_buy_{item['id']}"))
                 productKeyboard.add(types.InlineKeyboardButton("Назад", callback_data="products"))
-                bot.edit_message_caption(f'**{item["name"]}** \n {item["description"]} \n **Цена:** {item["price"]} руб.', call.message.chat.id, call.message.message_id, parse_mode='Markdown', reply_markup=productKeyboard)
+                bot.edit_message_media(chat_id=call.message.chat.id, message_id=call.message.message_id, media=types.InputMediaPhoto(item['image']), reply_markup=productKeyboard)
+                bot.edit_message_caption(chat_id=call.message.chat.id, message_id=call.message.message_id, caption=f"{item['name']}\n\n{item['description']}\n\nЦена: {item['price']} руб.",reply_markup=productKeyboard)
                 break
 
     if call.data.startswith("confirm_buy_"):
+        for customer in customers:
+            if customer['user_id'] == call.from_user.id and customer['status'] == 'pending':
+                bot.send_message(call.message.chat.id, "У вас уже есть незавершенный заказ.")
+                return
+
+        orderNumber = randint(100000, 999999)
         item_id = int(call.data.split("_")[2])
         for item in shopitems:
             if item['id'] == item_id:
-                # Здесь логика покупки, например, отправка сообщения о покупке
-                bot.send_message(call.message.chat.id, f"Вы купили {item['name']} за {item['price']} руб.")
-                break
+             
+             customers.append({
+                      "user_id": call.from_user.id,
+                      "product_id": item['id'],
+                      "status": "pending",
+                      "order_number": orderNumber
+                 })
+             with open(os.path.join(path, 'customers.json'), 'w', encoding='utf-8-sig') as f:
+                json.dump(customers, f, ensure_ascii=False, indent=4)
+             try:
+                  bot.delete_message(call.message.chat.id, call.message.message_id)  
+                  bot.send_message(chat_id=call.message.chat.id,text=str.format(settings['product_purchase_message'], product_name=item['name']))
+             except Exception as e:
+                                bot.send_message(chat_id=call.message.chat.id,text=str.format(settings['product_purchase_message'], product_name=item['name']))
 
+                              
+        
 
     if call.data == "back_to_welcome":
         start(call.message)
         bot.delete_message(call.message.chat.id, call.message.message_id)
         
+    if call.data.startswith("complete_order_"):
+        order_number = int(call.data.split("_")[2])
+        for customer in customers:
+            if customer['order_number'] == order_number:
+                customer['status'] = 'completed'
+                with open(os.path.join(path, 'customers.json'), 'w', encoding='utf-8-sig') as f:
+                    json.dump(customers, f, ensure_ascii=False, indent=4)
 
+                bot.edit_message_reply_markup(chat_id=call.message.chat.id, message_id=call.message.message_id, reply_markup=None)
+                bot.send_message(chat_id=customer['user_id'], text=f"Заказ {order_number} завершён! \n По всем вопросам, касающихся заказа и/или его выполнения обращайтесь в поддержку @Flaemee \n Спасибо за покупку!💖")
+                break
        
+    if call.data.startswith("edit_"):
+        item = find_product_by_id(int(call.data.split("_")[1]))
+        if item:
+            editKeyboard = types.InlineKeyboardMarkup(row_width=1)
+            for key in dataname.keys():
+                editKeyboard.add(types.InlineKeyboardButton(dataname[key], callback_data=f"edit_{key}_{item['id']}"))
+            editKeyboard.add(types.InlineKeyboardButton("Назад", callback_data="editproduct"))
+            bot.edit_message_text(chat_id=call.message.chat.id, message_id=call.message.message_id, text=f"Вы редактируете товар {item['name']}",reply_markup=editKeyboard)
 
-
+    if call.data.startswith("edit_") and len(call.data.split("_")) == 3:
+        item = find_product_by_id(int(call.data.split("_")[3]))
+        if item:
+            key = call.data.split("_")[1]
+            user_states[call.from_user.id] = {'stage': f'edit_{key}', 'item_id': item['id']}
+            bot.send_message(call.message.chat.id, f"Введите новое значение для {dataname[key]} (текущее значение: {item[key]})")
+            @bot.message_handler(func=lambda m: m.text and not m.text.startswith('/'))
+            def handle_edit_product(message):
+                state = user_states.get(message.from_user.id)
+                if state and state['stage'].startswith('edit_'):
+                    key = state['stage'].split("_")[1]
+                    item_id = state['item_id']
+                    item = find_product_by_id(item_id)
+                    if item:
+                        if key == 'price':
+                            try:
+                                item[key] = int(message.text.strip())
+                            except ValueError:
+                                bot.reply_to(message, "Цена должна быть числом. Попробуйте снова.")
+                                return
+                        else:
+                            item[key] = message.text.strip()
+                        with open(os.path.join(path, 'ShopItems.json'), 'w', encoding='utf-8-sig') as f:
+                            json.dump(shopitems, f, ensure_ascii=False, indent=4)
+                        bot.reply_to(message, f"{dataname[key]} товара '{item['name']}' успешно обновлено!")
+                        del user_states[message.from_user.id]  # Очистить состояние
+    
 def check_subscription(user_id):
     try:
-        member = bot.get_chat_member(-1002389669978, user_id)
+        member = bot.get_chat_member(settings['subscribecribed_channels'], user_id)
         return member.status in ['creator', 'administrator', 'member']
     except Exception as e:
         print(f"Ошибка при проверке подписки: {e}")
@@ -93,47 +176,65 @@ def start(message):
 @bot.message_handler(commands=['newproduct'])
 def new_product_create(message):
     user_id = message.from_user.id
-    if bot.get_chat_member(-1002389669978, user_id).status not in ['creator', 'administrator']:
+    if bot.get_chat_member(settings['subscribecribed_channels'], user_id).status not in ['creator', 'administrator']:
         bot.reply_to(message, "У вас нет прав для добавления товаров.")
         return
     user_states[user_id] = {'stage': 'name', 'new_product': {}}
     bot.send_message(message.chat.id, "Введите название нового товара")
 
+
+new_product = {}
+
 @bot.message_handler(func=lambda m: m.text and not m.text.startswith('/'))
 def handle_new_product_steps(message):
     user_id = message.from_user.id
+    
+    for customer in customers:
+        if customer['user_id'] == user_id and customer['status'] == 'pending':
+            confrimKeyboard = types.InlineKeyboardMarkup()
+            confrimKeyboard.add(types.InlineKeyboardButton("Выдача товара завершена", callback_data=f"complete_order_{customer['order_number']}"))
+            bot.send_message(chat_id=settings['work_channel'],text=str.format(settings['workers_purchase_notification'],user=f'@{message.from_user.username}',product_name=find_product_by_id(customer['product_id'])['name'],data=message.text),reply_markup=confrimKeyboard)
+            bot.reply_to(message, "Спасибо за предоставленную информацию! Ваш заказ будет обработан в ближайшее время.")
+            return
     if user_id not in user_states:
         return  # Игнорировать, если нет активного состояния
     
     state = user_states[user_id]
     stage = state['stage']
-    new_product = {}
+    
     
     if stage == 'name':
+        new_product.clear()
         new_product['name'] = message.text.strip()
+        print(new_product['name'])
         state['stage'] = 'price'
         bot.send_message(message.chat.id, "Введите цену нового товара")
+        print(new_product)
     elif stage == 'price':
         try:
             new_product['price'] = int(message.text.strip())
             state['stage'] = 'description'
             bot.send_message(message.chat.id, "Введите описание нового товара")
+            print(new_product)
         except ValueError:
             bot.reply_to(message, "Цена должна быть числом. Попробуйте снова.")
     elif stage == 'description':
         new_product['description'] = message.text.strip()
         state['stage'] = 'image'
         bot.send_message(message.chat.id, "Отправьте изображение товара или введите URL")
+        print(new_product)
     elif stage == 'image':
         if message.content_type == 'photo':
             file_info = bot.get_file(message.photo[-1].file_id)
             file_url = f"https://api.telegram.org/file/bot{settings['token']}/{file_info.file_path}"
             new_product['image'] = file_url
+            print(new_product)
         else:
             new_product['image'] = message.text.strip()
         
         # Добавление товара
         new_id = max(item['id'] for item in shopitems) + 1 if shopitems else 1
+        print(f"Пользователь {message.from_user.id} добавляет товар: {new_product}")
         new_item = {
             'id': new_id,
             'name': new_product['name'],
@@ -193,7 +294,7 @@ def handle_photo(message):
 @bot.message_handler(commands=['removeproduct'])
 def remove_product(message):
     user_id = message.from_user.id
-    if bot.get_chat_member(-1002389669978, user_id).status not in ['creator', 'administrator']:
+    if bot.get_chat_member(settings['subscribecribed_channels'], user_id).status not in ['creator', 'administrator']:
         bot.reply_to(message, "У вас нет прав для удаления товаров.")
         return
     user_states[user_id] = {'stage': 'remove'}
@@ -209,7 +310,29 @@ def remove_product(message):
                 del user_states[message.from_user.id]  # Очистить состояние
                 return
     
+@bot.message_handler(commands=['orders'])
+def view_orders(message):
+    if message.chat.id != settings['work_channel'] or not bot.get_chat_member(settings['subscribecribed_channels'], message.from_user.id).status in ['creator', 'administrator']:
+        bot.reply_to(message, "У вас нет доступа к этой команде.")
+        return
+    orderButtons = types.InlineKeyboardMarkup(row_width=2)
+    for order in customers:
+        orderButton = types.InlineKeyboardButton(f"Заказ {order['order_number']}", callback_data=f'view_order_{order["order_number"]}')
+        orderButtons.add(orderButton)
+    bot.send_message(message.chat.id, "Выберите заказ для просмотра:", reply_markup=orderButtons)
 
+@bot.message_handler(commands=['editproduct'])
+def edit_product(message):
+    user_id = message.from_user.id
+    if bot.get_chat_member(settings['subscribecribed_channels'], user_id).status not in ['creator', 'administrator']:
+        bot.reply_to(message, "У вас нет прав для редактирования товаров.")
+        return
+    editButtons = types.InlineKeyboardMarkup()
+    for item in shopitems:
+        editButtons.add(types.InlineKeyboardButton(f"{item['name']}", callback_data=f"edit_{item['id']}"))
+    bot.send_message(message.chat.id, f"Выберите товар для редактирования:", reply_markup=editButtons)
+   
+    
 bot.infinity_polling(
     allowed_updates=['message', 'callback_query']
 )
